+++ bee-block/src/payload/milestone/option/mod.rs
@@ -144,11 +144,14 @@ pub mod dto {
     use serde::{Deserialize, Serialize, Serializer};
     use serde_json::Value;
 
-    pub use self::{parameters::dto::ParametersMilestoneOptionDto, receipt::dto::ReceiptMilestoneOptionDto};
+    pub use self::{
+        parameters::dto::ParametersMilestoneOptionDto,
+        receipt::dto::{MigratedFundsEntryDto, ReceiptMilestoneOptionDto},
+    };
     use super::*;
     use crate::error::dto::DtoError;
 
-    #[derive(Clone, Debug, Eq, PartialEq)]
+    #[derive(Clone, Debug, Eq, PartialEq, From)]
     pub enum MilestoneOptionDto {
         /// A receipt milestone option.
         Receipt(ReceiptMilestoneOptionDto),