+++ bee-block/src/payload/transaction/essence/mod.rs
@@ -50,7 +50,7 @@ pub mod dto {
     use crate::{error::dto::DtoError, protocol::ProtocolParameters};
 
     /// Describes all the different essence types.
-    #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+    #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, From)]
     #[serde(untagged)]
     pub enum TransactionEssenceDto {
         Regular(RegularTransactionEssenceDto),