+++ bee-block/src/signature/mod.rs
@@ -46,7 +46,7 @@ pub mod dto {
     use crate::error::dto::DtoError;
 
     /// Describes all the different signature types.
-    #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+    #[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize, From)]
     #[serde(untagged)]
     pub enum SignatureDto {
         Ed25519(Ed25519SignatureDto),